docs: storage mock
