+++ docs/tsconfig.json
@@ -14,6 +14,9 @@
       "@aws-sdk/node-http-handler": [
         "./node_modules/@aws-sdk/node-http-handler/dist/es",
         "./node_modules/@aws-sdk/node-http-handler"
+      ],
+      "aws-amplify/storage": [
+        "storageMock.ts"
       ]
     },
     "allowJs": true,
